Fixed heat interface rendering
hr_version change:
Hr version was getting initialized with x value from normal res, so the sprite wasn't loaded correctly.
By replacing the normal res with the hr version it can be ensured that only properties from the hr version are loaded.
This matches the behaviour of the game, it also replaces the normal version with the hr version when loading.