Fixed ArtilleryTurret rendering
The filename_selector was only being set on the low res version. By moving it out of the lua, it can be set for hr and sr version and becomes less of a hack.
Note: Artillery turret shadows currently overlap due to me including all layers instead of only non-shadow layers.